Added workout get progress function to get the progress of a workout in, before or after progress

+++ Gym App.Infrastructure/DTOs/Workout/WorkoutSetProgressUpdateDTO.cs
@@ -6,7 +6,7 @@ namespace Gym_App.Infrastructure.DTOs.Workout
         public Guid SetId { get; set; }
         public bool IsCompleted { get; set; }
         public required int ActualReps { get; set; }
-        public required int ActualWeight { get; set; }
+        public required float ActualWeight { get; set; }
         public int KCaloriesBurned {  get; set; }
         public string? Notes { get; set; }
     }
